refact: move domain models to :model
