+++ app/build.gradle
@@ -37,5 +37,5 @@ android {
 dependencies {
     implementation project(':ui')
     implementation project(':ui-today')
-    implementation implementations.material
+    implementation libraries.material
 }
