+++ calculator/src/test/java/com/jeanbarrossilva/h2o/calculator/WaterIntakeCalculatorTests.kt
@@ -1,7 +1,7 @@
 package com.jeanbarrossilva.h2o.calculator
 
-import com.jeanbarrossilva.h2o.drinker.Age
-import com.jeanbarrossilva.h2o.drinker.Drinker
+import com.jeanbarrossilva.h2o.model.drinker.Age
+import com.jeanbarrossilva.h2o.model.drinker.Drinker
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
@@ -13,7 +13,7 @@ class WaterIntakeCalculatorTests {
     }
 
     @Test
-    fun `GIVEN a weight of 120 kg and an age of 35 WHEN calculating the goal THEN it's 4_200 mL`() {
+    fun `GIVEN a weight of 120 kg and an age of 35 WHEN calculating the goal THEN it's 4,200 mL`() {
         val drinker = Drinker("Carlos", Age of 35, weightInKilograms = 120)
         assertEquals(4_200, WaterIntakeGoalCalculator.calculate(drinker))
     }
