+++ logger/src/main/java/com/jeanbarrossilva/h2o/logger/InMemoryLogger.kt
@@ -1,13 +1,13 @@
 package com.jeanbarrossilva.h2o.logger
 
+import com.jeanbarrossilva.h2o.logger.moment.MomentProvider
 import com.jeanbarrossilva.h2o.model.intake.Intake
 import com.jeanbarrossilva.h2o.model.intake.IntakeLog
-import com.jeanbarrossilva.h2o.logger.time.MomentProvider
 
 data class InMemoryLogger(private val momentProvider: MomentProvider): Logger {
     private val entries = mutableListOf<IntakeLog>()
 
-    private val time
+    private val moment
         get() = momentProvider.provide()
 
     override suspend fun getLogs(): List<IntakeLog> {
@@ -21,7 +21,7 @@ data class InMemoryLogger(private val momentProvider: MomentProvider): Logger {
     }
 
     override suspend fun log(intake: Intake): IntakeLog {
-        return IntakeLog(time, intake).also {
+        return IntakeLog(moment, intake).also {
             entries.add(it)
         }
     }
