+++ logger/src/main/java/com/jeanbarrossilva/h2o/logger/moment/DefaultMomentProvider.kt
@@ -1,8 +1,8 @@
-package com.jeanbarrossilva.h2o.logger.time
+package com.jeanbarrossilva.h2o.logger.moment
 
 import java.time.LocalDateTime
 
-object DefaultMomentProvider: MomentProvider {
+class DefaultMomentProvider: MomentProvider {
     override fun provide(): LocalDateTime {
         return LocalDateTime.now()
     }
