+++ ui-today/build.gradle
@@ -46,8 +46,8 @@ dependencies {
     implementation project(':model')
     implementation project(':preferences')
     implementation project(':ui')
-    implementation implementations.compose.material
-    implementation implementations.constraintlayout.compose
-    implementation implementations.compose.material_icons_extended
-    implementation implementations.mpandroidchart
+    implementation libraries.compose.material
+    implementation libraries.constraintlayout.compose
+    implementation libraries.compose.material_icons_extended
+    implementation libraries.mpandroidchart
 }
